// webui/api/models/MachineRuntimeInfoDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MachineRuntimeInfoDTO = {
    timestamp?: string;
    serverId?: number | null;
    cpuUsage?: number | null;
    memoryUsage?: number | null;
    swapUsage?: number | null;
    diskUsage?: number | null;
};